Move StaticValue to :model-core
