+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/DefaultTaskOutputFilePropertyRegistration.java
@@ -18,6 +18,7 @@ package org.gradle.api.internal.tasks;
 
 import org.gradle.api.tasks.TaskOutputFilePropertyBuilder;
 import org.gradle.internal.properties.OutputFilePropertyType;
+import org.gradle.internal.properties.StaticValue;
 
 public class DefaultTaskOutputFilePropertyRegistration extends AbstractTaskFilePropertyRegistration implements TaskOutputFilePropertyRegistration {
     private final OutputFilePropertyType outputFilePropertyType;
